kubectl: move printObjectSpecificMessage in factory
