+++ pkg/kubectl/cmd/cmd_test.go
@@ -334,6 +334,7 @@ func NewAPIFactory() (*cmdutil.Factory, *testFactory, runtime.Codec) {
 	f.ProtocolsForObject = rf.ProtocolsForObject
 	f.LabelsForObject = rf.LabelsForObject
 	f.CanBeExposed = rf.CanBeExposed
+	f.PrintObjectSpecificMessage = rf.PrintObjectSpecificMessage
 	return f, t, testapi.Default.Codec()
 }
